Create scopes module
